requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
Django==2.1
djangorestframework==3.8.2
docopt==0.6.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
paramiko==2.4.1
pkg-resources==0.0.0
prompt-toolkit==2.0.4
pyasn1==0.4.4
pycparser==2.18
pyflakes==2.0.0
Pygments==2.2.0
PyNaCl==1.2.1
pytz==2018.5
pyvim==2.0.22
pyvmomi==6.7.0
requests==2.19.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
wget==3.2
